Carousel component fixed
Fixed a large portion of css issues. Changed width of the card because it felt too big when at 80%.
Solved why the video player was so poorly sized.
Added functionality to arrow buttons to scroll to the next item in the carousel.
Didn't remove the national geographic giraffe or the randomizer image.